requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
python-dotenv==1.2.1
PyMySQL==1.1.2           # Alternatively use mysqlclient with slight config change
Flask-Login==0.6.3
Werkzeug==3.1.0
Pillow==12.0.0               # for image handling later
